// admin-clients/libs/cpanel/promoters/season-tickets/channels/feature/src/lib/models/season-channel-sale-status.enum.ts
export enum SeasonChannelSaleStatus {
    pendingRelationship = 'PENDING_RELATIONSHIP',  // PENDIENTE_RELACION
    planned = 'PLANNED',                            // PLANIFICACION
    inProgramming = 'IN_PROGRAMMING',              // PROGRAMACION
    salePending = 'SALE_PENDING',                  // PENDIENTE_VENTA
    sale = 'SALE',                                  // EN_VENTA
    saleCancelled = 'SALE_CANCELLED',              // VENTA_CANCELADA
    releaseFinished = 'RELEASE_FINISHED',          // PUBLICACION_FINALIZADA
    releaseCancelled = 'RELEASE_CANCELLED',        // PUBLICACION_CANCELADA
    cancelled = 'CANCELLED',                        // CANCELADO
    notAccomplished = 'NOT_ACCOMPLISHED',          // NO_REALIZADO
    rejected = 'REJECTED'                          // RECHAZADO
}